pyproject.toml
[project]
name = "mcp-on-cloudrun"
version = "0.1.0"
description = "Example of deploying a MCP server on Cloud Run"
requires-python = ">=3.10"
dependencies = [
    "aiohttp==3.12.13",
    "anthropic>=0.54.0",
    "fastapi>=0.115.13",
    "fastmcp==2.6.1",
    "openai>=1.88.0",
    "python-dotenv==1.1.0",
]
